% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{add_zoom_fun}
\alias{add_zoom_fun}
\title{Función interna para hacer zoom}
\usage{
add_zoom_fun(
  .plot,
  coor_xmin = -100.29,
  coor_xmax = -93.71,
  coor_ymin = 15.6,
  coor_ymax = 20.21,
  ub_xmin = -119.5,
  ub_xmax = -106,
  ub_ymin = 13.5,
  ub_ymax = 22.8
)
}
\arguments{
\item{.plot}{plots}
}
\value{
sf object
}
\description{
...
}
