% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-segpoint.R
\name{draw_key_segpoint}
\alias{draw_key_segpoint}
\title{Custom Key Drawing Function for GeomSegPoint (draw_key_segpoint)}
\usage{
draw_key_segpoint(data, params, size)
}
\arguments{
\item{data}{A data frame containing information about the legend key to 
be drawn.}

\item{params}{A list of parameters specifying aesthetics and properties 
of the geom.}

\item{size}{The size of the legend key.}
}
\value{
A \code{grid::grobTree} representing the legend key for the
\code{\link{GeomSegPoint}} geom.
}
\description{
Custom key drawing function for the \code{\link{GeomSegPoint}} geom.
This function is responsible for drawing the legend key for the segments 
and points in the plot.
}
