% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_geoms.R
\name{formt_text}
\alias{formt_text}
\title{Función para determinar qupe función usar}
\usage{
formt_text(markdown = F, ...)
}
\arguments{
\item{markdown}{booleano para determinar si usar la función de base o la
de {{ggtext}}.}

\item{...}{los elementos correspondientes al formato de element_text.}
}
\value{
una función con base en la elección
}
\description{
Esa función determina si usar formato markdawn o texto normal.
}
