% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-segpoint.R
\name{geom_segpoint}
\alias{geom_segpoint}
\title{Geometric Segmented Points (geom_segpoint)}
\usage{
geom_segpoint(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  arrow = NULL,
  width = 0.4,
  arrow.fill = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{A list specifying how variables in the data are mapped 
to aesthetic properties (e.g., x, y, color, size) of the geom.}

\item{data}{A data frame or tibble containing the data to be visualized.}

\item{stat}{The statistical transformation to be applied to the data 
before plotting.}

\item{position}{The method used to position the data points on the plot.}

\item{...}{Additional parameters and arguments to be passed to other 
functions or layers.}

\item{arrow}{Optional arrow specification for the segments, allowing you 
to customize the arrow's appearance.}

\item{width}{The width of the segments drawn from the data points. Default 
is 0.4.}

\item{arrow.fill}{The fill color of the arrow. If not provided, it defaults 
to the color specified in the plot.}

\item{na.rm}{A logical value indicating whether to remove missing values 
from the data before plotting.}

\item{show.legend}{A logical value indicating whether to show the legend 
for this layer.}

\item{inherit.aes}{A logical value indicating whether to inherit the 
aesthetics from the parent ggplot2 object.}
}
\value{
A ggplot2 layer representing the geom \code{\link{GeomSegPoint}}.
}
\description{
Add segmented points to a ggplot2 plot, combining segments and points to
represent data points. This geom is particularly useful for adding marks 
over the geom_col to provide additional information without the need for 
a extra bar.
}
\examples{
# Basic example
set.seed(1234)
db <- data.frame(category = 2001:2020,
                 val1 = rnorm(20, 100, 10),
                 val2 = rnorm(20, 100, 10))
ggplot(data = db, aes(x = category, y = val1)) +
  geom_col() +
  geom_segpoint(aes(y = val2))

}
\seealso{
\code{\link{GeomSegPoint}} for the geom object and additional details.
}
