% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_geoms.R
\name{geom_text_bi}
\alias{geom_text_bi}
\title{Geom text para barras}
\usage{
geom_text_bi(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  parse = FALSE,
  nudge_x = 0,
  nudge_y = 0,
  check_overlap = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  position_hor = T,
  percent_change = 0.25,
  color_black = "grey15",
  color_light = "grey95"
)
}
\arguments{
\item{data}{input1}
}
\value{
Una nueva capa de texto para una gráfica
}
\description{
Esta función sirve para añadir fácilmente los valores para los barras de variación
en valores negativos y positivos. Sólo se esplican los parámetros que difieren de
geom_text() normal
}
