% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-segpoint.R
\docType{data}
\name{GeomSegPoint}
\alias{GeomSegPoint}
\title{GeomSegPoint object}
\format{
An object of class \code{GeomSegPoint} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 7.
}
\usage{
GeomSegPoint
}
\description{
Custom ggproto object that defines the properties and behavior of the
\code{geom_segpoint} function. The \code{GeomSegPoint} object inherits 
from the \code{Geom} base class and is used to create the segments and 
points for the \code{geom_segpoint} geom.
}
\section{Fields}{

\describe{
\item{\code{required_aes}}{A character vector specifying the aesthetics required 
by this geom. In this case, "x" and "y" are required to represent the 
data points' positions.}

\item{\code{non_missing_aes}}{A character vector specifying other aesthetics that 
can be used with this geom, even if they have missing values.}

\item{\code{default_aes}}{Aesthetic mappings that are applied by default to the 
geom, in case the user does not explicitly specify them.}
}}

\seealso{
\code{\link[ggplot2]{Geom}} for details about the base class.
}
\keyword{datasets}
