% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_geoms.R
\docType{data}
\name{GeomTextBi}
\alias{GeomTextBi}
\title{Ggproto del goeom text para barras}
\format{
An object of class \code{GeomTextBi} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomTextBi
}
\arguments{
\item{data}{input1}
}
\value{
Una nueva capa de texto para una gráfica
}
\description{
Esta función sirve para añadir fácilmente los valores para los barras de variación
en valores negativos y positivos. Sólo se esplican los parámetros que difieren de
geom_text() normal
}
\keyword{datasets}
