% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{ggplot_add.my_zoom}
\alias{ggplot_add.my_zoom}
\title{Añadir método para hacer el zoom}
\usage{
\method{ggplot_add}{my_zoom}(object, plot, object_name)
}
\arguments{
\item{..x}{}
}
\value{
sf object
}
\description{
...
}
