% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{jmr_pal_manual}
\alias{jmr_pal_manual}
\title{Color interpolacion de paletas manual}
\usage{
jmr_pal_manual(n, palette = "general", reverse = FALSE, ...)
}
\arguments{
\item{n}{número de colores.}

\item{palette}{Determina qué paleta tomar.}

\item{reverse}{Determina el orden de la paleta.}

\item{...}{Demás elementos correspondientes a grDevices::colorRampPalette()}
}
\value{
regresa una función con un argumento integer (el número de colores)
y un vector de colores.
}
\description{
Esta función genera nuevos colores cuando es necesario y determina orden.
}
