% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{make_limit}
\alias{make_limit}
\title{Límites determinados por geometries}
\usage{
make_limit(.x, lon_buff = 0.2, lat_buff = 0.2)
}
\arguments{
\item{.x}{}
}
\value{
layer con límites
}
\description{
...
}
