% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_geoms.R
\name{mi_tema}
\alias{mi_tema}
\title{Tema inicial con formato general}
\usage{
mi_tema(...)
}
\arguments{
\item{...}{cambios en el tema}
}
\value{
Tema para gráficas de ggplot
}
\description{
Esta resume las funciones de estilo utilizadas en mis gráficas
}
