% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_geoms.R
\name{mi_tema_html}
\alias{mi_tema_html}
\title{Tema inicial con formato html}
\usage{
mi_tema_html(...)
}
\arguments{
\item{...}{cambios en el tema}
}
\value{
Tema en html para gráficas de ggplot
}
\description{
Esta resume las funciones de estilo utilizadas en mis gráficas
con estilo html
}
