% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_objects.R
\name{month_to_number}
\alias{month_to_number}
\title{Transformar mes a número}
\usage{
month_to_number(mes)
}
\arguments{
\item{...}{cambios en el tema}
}
\value{
Tema en html para gráficas de ggplot
}
\description{
Esta resume las funciones de estilo utilizadas en mis gráficas
con estilo html
}
