% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_objects.R
\name{number_to_month}
\alias{number_to_month}
\title{Transformar número a mes}
\usage{
number_to_month(mes)
}
\arguments{
\item{...}{cambios en el tema}
}
\value{
Tema en html para gráficas de ggplot
}
\description{
Esta resume las funciones de estilo utilizadas en mis gráficas
con estilo html
}
