% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\docType{data}
\name{paletas_mcci}
\alias{paletas_mcci}
\title{Paletas de colores de MCCI}
\format{
list
}
\source{
Creación de comunicación de MCCI
}
\usage{
paletas_mcci
}
\description{
Es una lista de las paletas de MCCI
}
\keyword{datasets}
