% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-excel-mrow.R
\name{read_excel_mrow}
\alias{read_excel_mrow}
\title{Read Excel File with Multi-Row Titles (read_excel_mul)}
\usage{
read_excel_mrow(path, range_title, sheet = NULL, ...)
}
\arguments{
\item{path}{The path to the Excel file to be read.}

\item{range_title}{A numeric vector of length 2 indicating the range of rows that contain the title information.
The titles should be displayed in these rows, and the data starts from the row following \code{range_title[2]}.}

\item{sheet}{The name or index of the sheet in the Excel file to read. If \code{NULL}, the default sheet is used.}

\item{...}{Additional arguments to be passed to \code{\link[readxl::read_excel]{readxl::read_excel}}.}
}
\value{
A data frame containing the data from the Excel file with multi-row titles.
}
\description{
Read an Excel file containing data with multi-row titles and extract the data into a data frame. The titles
are assumed to be displayed consecutive rows, and the data is read from the row following the titles.
}
