% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{scale_color_mcci}
\alias{scale_color_mcci}
\title{Función para añadir colores de MCCI a gráfica}
\usage{
scale_color_mcci(palette = "general", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Determina qué paleta tomar. Puede ser general, bivariado1, 
bivariado2, variado y multiple.}

\item{discrete}{Determina si los datos son discretos o continuos.}

\item{reverse}{Determina el orden de la paleta.}

\item{...}{Demás elementos correspondientes a scale_color_gradientn()}
}
\value{
regresa una función con un argumento integer (el número de colores)
y un vector de colores.
}
\description{
Esta función añade la paleta de colores a la gráfica de ggplot.
}
