% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{set_mytheme}
\alias{set_mytheme}
\title{Set Custom Plot Theme}
\usage{
set_mytheme(geoms = c("bar", "col", "area", "point", "boxplot", "bin"), ...)
}
\arguments{
\item{geoms}{A character vector specifying the ggplot2 geometric objects for which to set default styles. Defaults to `c("bar", "col", "area", "point", "boxplot", "histogram")`.}

\item{...}{Additional arguments passed to `theme_jmr`, allowing customization of the plot theme.}
}
\value{
This function is used for its side effects of setting theme and display options. It does not return a value.
}
\description{
This function configures a customized theme for ggplot2 plots, setting locale options, disabling scientific notation, and applying specified theme settings to specified ggplot2 geometric objects.
}
\details{
This function sets a customized ggplot2 theme by applying `theme_jmr` globally and configuring color and fill palettes based on `paletas_jmr$generalextend`. It also sets locale options if running on a macOS system, disables scientific notation, and updates the default aesthetics of specified ggplot2 geoms to use colors from the `paletas_jmr` palette with adjusted transparency levels.
}
