% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{theme_jmr}
\alias{theme_jmr}
\title{Tema inicial con formato propio}
\usage{
theme_jmr(..., text_resize = 1, markdown = F)
}
\arguments{
\item{...}{cambios en el tema}
}
\value{
Tema para gráficas de ggplot
}
\description{
Esta resume las funciones de estilo utilizadas en mis gráficas
}
