% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{theme_maps_jmr}
\alias{theme_maps_jmr}
\title{Tema para mapas}
\usage{
theme_maps_jmr(...)
}
\arguments{
\item{...}{cambios en el tema}
}
\value{
Tema en html para gráficas de ggplot
}
\description{
Esta resume las funciones de estilo utilizadas en mis gráficas
con estilo html
}
